React Portfolio Project
This project was coded in react using typescript. First to be able to view and use it, you must have a node.js version 16 or higher. If you are unsure of how to use react Programming with Mosh on YouTube has an excellent tutorial to get you started. I will have the link in this description if you needed it. If you were you edit this project, you would go into the src folder, and then the components folder. Within the components folder, there is the App.tsx which acts as the main file for the program. There is also NavBar.tsx, and Footer.tsx that is called and used for every state of the project. In between the two should be another .tsx file called, whether it be Home, Research or Accolades. If you wanted to add, delete, or edit any of the data displayed, you would want to do it within those .tsx files. If something seems out of line or incorrectly sized, you would want to use the .css for the appropriate  area. I believe there is still room for improvement, but with the allotted time I was given and able to work, this was the best I code. 

Link to React Tutorial: https://www.youtube.com/watch?v=SqcY0GlETPk